changes implemented in a bit
