+++ .gitignore
@@ -78,3 +78,10 @@ build
 # Ignore build files
 public
 
+
+/public/packs
+/public/packs-test
+/node_modules
+/yarn-error.log
+yarn-debug.log*
+.yarn-integrity
