+++ app/javascript/App.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { Routes, Route } from "react-router-dom";
+import RandomMessage from "./src/components/Greeting.js";
+
+const App = () => {
+  return (
+    <React.Fragment>
+      <Routes>
+        <Route path="/" element={<RandomMessage />} />
+      </Routes>
+    </React.Fragment>
+  );
+};
+
+export default App;
