+++ app/javascript/application.js
@@ -1,1 +1,20 @@
-import './react/src/index'
+import "@hotwired/turbo-rails";
+import "./controllers";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Provider from 'react-redux';
+import App from './App';
+import { BrowserRouter as Router } from 'react-router-dom';
+import store from './src/redux/store';
+
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <Router>
+        <App />
+      </Router>
+    </Provider>
+  </React.StrictMode>
+);
