+++ package.json
@@ -5,6 +5,7 @@
     "@babel/preset-env": "^7.22.4",
     "@hotwired/stimulus": "^3.2.2",
     "@hotwired/turbo-rails": "^7.3.0",
+    "@rails/webpacker": "5.4.4",
     "@reduxjs/toolkit": "^2.1.0",
     "axios": "^1.4.0",
     "babel-loader": "^9.1.2",
@@ -19,7 +20,7 @@
     "redux-thunk": "^2.4.2",
     "sort-by": "^1.2.0",
     "webpack": "^5.90.1",
-    "webpack-cli": "^5.1.4"
+    "webpack-cli": "5.1.4"
   },
   "scripts": {
     "build": "webpack --config webpack.config.js"
